More cleanup
I'm putting this up in a presentable form it will
hopefully be interesting to someone. I derived it in a different way
than Doug because I didn't understand what in tarnation he was doing but
my approach has an issue with whether the points are in the expected
order (clockwise I think) and will break if the vectors are ordered
wrong. In thinking how to fix that, I finally came to understand what
Doug was doing I think, and am going to rework this to follow his method
and also be robust to airspeed vector ordering.